requirements.txt
# Core ML and Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0

# MLflow
mlflow>=2.8.0
mlflow-skinny>=2.8.0

# API and Web
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Monitoring and Visualization
plotly>=5.13.0
dash>=2.9.0
dash-bootstrap-components>=1.4.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0

# Additional dependencies
torch>=2.2.0
hyperopt>=0.2.7
seaborn>=0.12.0
matplotlib>=3.7.0
psycopg2-binary>=2.9.0  # For PostgreSQL support 